requirements.txt
# requirements.txt
gradio
gspread
google-auth
transformers
torch
google-generativeai
pydantic